chore: wait for db initialize
