+++ package.json
@@ -35,6 +35,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
+    "driver.js": "^1.3.6",
     "drizzle-orm": "^0.44.2",
     "lucide-vue-next": "^0.523.0",
     "nprogress": "^0.2.0",
