+++ src/pages/index.vue
@@ -1,12 +1,19 @@
 <script setup lang="ts">
+import { until } from '@vueuse/core'
+import { storeToRefs } from 'pinia'
 import { onMounted } from 'vue'
 import { useRouter } from 'vue-router'
+import { useDatabaseStore } from '~/stores/database'
 import { useRoomsStore } from '~/stores/rooms'
 
 const router = useRouter()
 const roomsStore = useRoomsStore()
+const dbStore = useDatabaseStore()
+const { db } = storeToRefs(dbStore)
 
 onMounted(async () => {
+  await until(db).toBeTruthy()
+
   // Initialize rooms
   const room = await roomsStore.initialize()
